+++ src/routes/MyPage/OrderList/OrderInfo.tsx
@@ -1,10 +1,8 @@
-import { useEffect, useState } from "react";
 import { IoIosArrowForward } from "react-icons/io";
-import { useInView } from "react-intersection-observer";
-import { Link, useLocation } from "react-router-dom";
+import { Link } from "react-router-dom";
 import styled from "styled-components";
 
-import { getOrderList } from "@/api/OrderApi";
+import useOrderList from "@/hooks/useOrderList";
 
 import OrderListBox from "./OrderListBox";
 
@@ -55,32 +53,7 @@ const orderStatus = [
 export default function OrderInfo() {
   // const [orderList, setOrderList] = useState<OrderListInfo[]>(mockData.body);
 
-  const location = useLocation();
-  const pathname = location.pathname.replace("/", "");
-  const isOrderListPage = pathname.includes("/order");
-
-  const [ref, inView] = useInView();
-  const [orderList, setOrderList] = useState<OrderListInfo[]>([]);
-  const [nextKey, setNextKey] = useState<number | null>(null);
-
-  const fetchOrderList = async () => {
-    const {
-      cursorRequest: { key },
-      body: data,
-    } = await getOrderList(nextKey, 5);
-    !isOrderListPage
-      ? setOrderList(data.splice(0, 4))
-      : setOrderList((prev) => [...prev, ...data]);
-    setNextKey(key);
-  };
-  useEffect(() => {
-    fetchOrderList();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  useEffect(() => {
-    if (inView && nextKey !== -1) fetchOrderList();
-  }, [inView]);
+  const { ref, isOrderListPage, orderList } = useOrderList();
 
   return (
     <>
